PR Fix: Move loading to App.vue
