+++ templates/vue/src/App.vue
@@ -1,8 +1,9 @@
 <template>
-  <div id="app">
+  <loading v-if="loading" style="height: 75vh;"></loading>
+  <div v-else id="app">
     <tapestry-app />
-    <router-view v-if="tapestryIsLoaded"></router-view>
-    <tapestry-sidebar v-if="tapestryIsLoaded" />
+    <router-view></router-view>
+    <tapestry-sidebar />
     <node-modal
       :node-id="nodeId"
       :modal-type="modalType"
@@ -13,14 +14,17 @@
 </template>
 
 <script>
-import { mapState } from "vuex"
+import { mapState, mapMutations } from "vuex"
 import NodeModal from "./components/NodeModal"
 import TapestryApp from "./components/TapestryApp"
 import TapestrySidebar from "./components/TapestrySidebar"
+import Loading from "./components/Loading"
+import client from "./services/TapestryAPI"
 
 export default {
   name: "app",
   components: {
+    Loading,
     NodeModal,
     TapestryApp,
     TapestrySidebar,
@@ -29,12 +33,20 @@ export default {
     return {
       modalType: "",
       nodeId: null,
+      loading: true,
     }
   },
   computed: {
-    ...mapState(["tapestryIsLoaded", "selectedNodeId"]),
+    ...mapState(["selectedNodeId"]),
   },
   mounted() {
+    window.addEventListener("click", this.recordAnalytics)
+    const data = [client.getTapestry(), client.getUserProgress()]
+    Promise.all(data).then(([dataset, progress]) => {
+      this.init({ dataset, progress })
+      this.loading = false
+    })
+
     this.$root.$on("add-node", to => {
       this.modalType = "add"
       this.nodeId = to
@@ -47,10 +59,22 @@ export default {
       this.$bvModal.show("node-modal")
     })
   },
+  beforeDestroy() {
+    window.removeEventListener("click", this.recordAnalytics)
+  },
   methods: {
+    ...mapMutations(["init"]),
     closeModal() {
       this.$bvModal.hide("node-modal")
       this.modalType = ""
+    },
+    recordAnalytics(evt) {
+      const x = evt.clientX + window.scrollLeft
+      const y = evt.clientY + window.scrollTop
+      client.recordAnalyticsEvent("user", "click", "screen", null, {
+        x: x,
+        y: y,
+      })
     },
   },
 }
