+++ templates/vue/src/components/TapestryApp.vue
@@ -1,6 +1,5 @@
 <template>
-  <loading v-if="loading" style="height: 75vh;"></loading>
-  <div v-else id="app-container">
+  <div id="app-container">
     <div class="toolbar">
       <tapestry-filter style="z-index: 10;" />
       <div class="slider-wrapper">
@@ -51,20 +50,17 @@
 <script>
 import DragSelect from "dragselect"
 import { mapMutations, mapState } from "vuex"
-import Loading from "@/components/Loading"
 import TapestryNode from "@/components/TapestryNode"
 import TapestryLink from "@/components/TapestryLink"
 import TapestryDepthSlider from "@/components/TapestryDepthSlider"
 import SettingsModalButton from "@/components/SettingsModalButton"
 import RootNodeButton from "@/components/RootNodeButton"
 import LockedTooltip from "@/components/LockedTooltip"
-import client from "@/services/TapestryAPI"
 import TapestryFilter from "@/components/TapestryFilter"
 import Helpers from "@/utils/Helpers"
 
 export default {
   components: {
-    Loading,
     TapestryNode,
     TapestryLink,
     TapestryDepthSlider,
@@ -81,14 +77,7 @@ export default {
     }
   },
   computed: {
-    ...mapState([
-      "nodes",
-      "links",
-      "selectedNodeId",
-      "tapestryIsLoaded",
-      "selection",
-      "settings",
-    ]),
+    ...mapState(["nodes", "links", "selectedNodeId", "selection", "settings"]),
     background() {
       return this.settings.backgroundUrl
     },
@@ -109,28 +98,10 @@ export default {
     },
   },
   mounted() {
-    window.addEventListener("click", this.recordAnalytics)
-
-    const data = [client.getTapestry(), client.getUserProgress()]
-    Promise.all(data).then(([dataset, progress]) => {
-      this.init({ dataset, progress })
-      this.loading = false
-      this.$nextTick(this.initializeDragSelect)
-    })
-  },
-  beforeDestroy() {
-    window.removeEventListener("click", this.recordAnalytics)
+    this.initializeDragSelect()
   },
   methods: {
-    ...mapMutations(["init", "select", "unselect", "clearSelection"]),
-    recordAnalytics(evt) {
-      const x = evt.clientX + window.scrollLeft
-      const y = evt.clientY + window.scrollTop
-      client.recordAnalyticsEvent("user", "click", "screen", null, {
-        x: x,
-        y: y,
-      })
-    },
+    ...mapMutations(["select", "unselect", "clearSelection"]),
     addRootNode() {
       this.$root.$emit("add-node", null)
     },
